+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sat Oct 01 06:50:08 MSK 2022
-sdk.dir=C\:\\Users\\Ivanov\\AppData\\Local\\Android\\Sdk
+#Sun Oct 02 22:18:51 MSK 2022
+sdk.dir=C\:\\Users\\admin\\AppData\\Local\\Android\\Sdk
